Refinements in MAXSAT, BinaryVectorSolution class now in binvec_solution, README updated
